+++ java/lib/src/main/java/com/truelayer/signing/Verifier.java
@@ -66,11 +66,13 @@ public abstract class Verifier {
      *
      * @param path - the request absolute path must not be null
      * @return the Verifier instance
-     * @throws IllegalArgumentException if the provided param is null
+     * @throws IllegalArgumentException if the provided param is null or invalid
      */
     public Verifier path(String path) {
         if (path == null)
             throw new IllegalArgumentException("the path must not be null");
+        if(!path.startsWith("/"))
+            throw new IllegalArgumentException("invalid path " + path + " must start with '/'");
         this.path = path;
         return this;
     }