+++ go/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.7
+* When verifying permit signed/verified path single trailing slash mismatches.
+
 ## 0.1.6
 * Added `Jku` method to `Signer` to allow different jku's endpoints.